requirements.txt
# FastAPI core
fastapi==0.110.0
uvicorn==0.29.0

# Database
pymongo==4.6.1

# Environment config
python-dotenv==1.0.1

# Security (Auth)
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-jose==3.3.0

# Data validation
pydantic==2.6.4

# Optional but recommended (for request parsing, forms, etc.)
python-multipart==0.0.9

# Logging & utilities
loguru==0.7.2

# Testing
pytest==8.1.1
httpx==0.27.0

# Async support (future-proofing)
anyio==4.3.0